Added endpoint for adding issues, added azure service to store images in azure blob storage, fixed model classes to get rid of exceptions

+++ src/main/java/com/pwr/jestsprawa/repositories/IssuesRepository.java
@@ -2,7 +2,6 @@ package com.pwr.jestsprawa.repositories;
 
 
 import com.pwr.jestsprawa.model.Issue;
-import com.pwr.jestsprawa.model.IssueDto;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
@@ -12,7 +11,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface IssuesRepository extends CrudRepository<Issue, Long> {
+public interface IssuesRepository extends CrudRepository<Issue, Integer> {
     @Query(value="SELECT i.* FROM app.issues AS i JOIN " +
             "(SELECT ss1.* FROM app.issuesstatuses ss1 LEFT JOIN app.issuesstatuses ss2 " +
             "ON ss1.issue_id = ss2.issue_id " +
